requirements.txt
pyautogen==0.2.16
openai==1.6.1
python-dotenv==1.0.0
docker==7.0.0
gitpython==3.1.40
autopep8==2.0.4
flake8==6.1.0
pytest==7.4.3
black==23.12.1